Include must-gather image name in ENV
The request for such feature came from a Bugzilla
for odf-must-gather.

We want the must-gather containers to be able to check the
image being used. This is useful in cases where we want to
deduce if the customer is using the image we ask them to.

While it is possible to call the k8s API internally (this is
how it is implemented currently for odf-mg), it would be much
efficient if the value is present inside the container as an ENV.

This patch sets an ENV var `MUST_GATHER_IMAGE` to reflect the
image source of the pods.

Signed-off-by: Niraj Yadav <niryadav@redhat.com>