`upgrade status`: improve tests for missing annotations
We discovered our existing test was triggering "weird" paths in the assessment code, because the input node was missing a MCO annotation while the node was ready (and it was missing only one annotation). While this is a possible scenario, it would likely need to be created manually by someone messing with annotations or the like. It did not represent the common scenario otherwise.

The typical scenario for missing annotations is a freshly provisioned node, before MCO has a chance to reconcile it. Such node is typically also not ready yet.