pkg/cli/admin/upgrade/: Add `oc adm upgrade recommend` subcommand
Add the backbones for the new `oc adm upgrade recommend` subcommand.

The logic is copied over from an existing package `upgrade` [1]. The
logic is to be changed over the upcoming iterations of development.

For the initial commit, the new `recommend` package contains the logic
from the [1] package as of the [2] commit. Only the logic regarding the
default output of the `oc adm upgrade` command and outputs regarding
available or available but not recommended updates was copied over.

This logic was placed behind a feature gate environment variable named
`OC_ENABLE_CMD_UPGRADE_RECOMMEND`. The new subcommand can be tested by
running:

```
$ OC_ENABLE_CMD_UPGRADE_RECOMMEND=true ./oc adm upgrade recommend
```

As most of the internal functions are not exported, the existing tests
were copied over as well for the existing used functions in the new
package `recommend`.

[1] github.com/openshift/oc/pkg/cli/admin/upgrade
[2] 93286c9260285c104259f0c50704c4b880c4fec3