Rename rsync'd ISO file to -output-name value
Updated unit-test to change the assertion on output filename to
assertion on the destination path.

Previously the filename was asserted against the RSyncOptions's
Destination field which contained the target filename and path.

Now the Destination field is a directory path. Because fs.FS
doesn't support write operations we cannot fake writing or renaming
the output file, which also blocks asserting the final output name
and path.