pkg/cli/admin/upgrade/recommend/examples: Update examples
Renaming 4.14.1-both-available-and-conditional to
4.14.1-all-recommended, because it has no Recommended!=True
conditionalUpdates.  And all-recommended seems like a more clear way
to express that, while "both-available-and-conditional" could sound
like "some are conditional and Recommended!=True".

Also bringing in some new examples from a 'launch 4.12.16 azure' Cluster
Bot install, followed by:

  $ oc adm upgrade channel stable-4.13

I collected the examples with:

  $ oc get clusterversion version -o yaml >pkg/cli/admin/upgrade/recommend/examples/4.12.16-...-cv.yaml

one early after the install, when the cluster-version operator was
slowly populating its cache of PromQL evaluations (later fixed in
4.12.43 [1]).  I've named the early-after-install
longest-not-recommended, because it has long hops that are not
recommended because risks have not been evaluated yet.  I've named the
while-after-install longest-recommended, because risks have been
evaluated.

[1]: https://issues.redhat.com/browse/OCPBUGS-22408?focusedId=23491059&page=com.atlassian.jira.plugin.system.issuetabpanels:comment-tabpanel#comment-23491059