pkg/cli/admin/upgrade/recommend: Show most-recent update options
We had been showing all the recommended updates in decreasing order
with --include-not-recommended to see all the
updates-with-assessed-risks in decreasing order.  But sometimes users
want to update to the longest-hop, even if there are known risks.  Or
they want to read about the assessed risks, in case there's something
they can do to their cluster to mitigate a currently-assessed risk
before kicking off the update.  This commit adjusts the output to
order roughly by release freshness For example, for a 4.y cluster in a
4.(y+1) channel:

  4.(y+1).tip
  4.(y+1).(tip-1)
  4.y.tip
  4.y.(tip-1)
  ...

Because users are more likely to care about 4.(y+1).tip, even if it
has assessed risks, than they are to care about 4.y.reallyOld, even if
it doesn't have assessed risks.

I'm showing two releases per z-stream at the moment, with
--show-outdated-releases to see all the results ("releases" to leave
the door open to having things like a single-arch and a multi-arch
target with the same version string).  And I'm dropping
--include-not-recommended, now that we are only filtering by age.

I'm also dropping "where are we now?" details, like Progressing=False,
because those are less relevant for folks who are trying to pick which
release they want to target next.  I still call out Failing=True and
Progressing=True, because 'oc adm upgrade ...' has client-side checks
for those ("the cluster is experiencing an error reconciling" and "the
cluster is already upgrading", with --allow-upgrade-with-warnings).
Folks who are concerned with how their cluster is currently doing can
run 'oc adm upgrade ...' or the tech-preview 'oc adm upgrade status ...').

I'm also dropping ReleaseAccepted, which is in a weird space after the
admin requests an update but before the cluster-version operator
accepts the retarget request.  Because 'oc adm upgrade' is the
subcommand used to request an update (--to and --to-image), I think
that's the appropriate place to surface that condition (and it already
does), so we don't need to surface that condition in the 'recommend'
subcommand (which is only relevant in helping the admin decide if and
how to request an update).