pkg/cli/admin/upgrade/recommend: Add --version option for specific releases
This allows us to drop the detailed, multi-line entries from the long
--show-outdated-releases output.  Now that form just gives us an
overview of releases with 'reason' slugs for any updates with assessed
issues.  Cluster administrators who see something in that overview
that they are interested in can make a subsequent call with
'--version VERSION' to get the long-form details on that particular
target.

I'm keeping the long-form output for release-exposed risks in the
default rendering though, because part of the motivation behind
d8cac253ac (pkg/cli/admin/upgrade/recommend: Show most-recent update
options, 2024-09-26, #1889) is to surface those longest-hop risk
messages for cluster administrators, in case the issue is something
like EvaluationFailed, which says more about the current cluster state
than it does about the health of the update.

Because the '--version 4.12.51' runs fail for some of our
ClusterVersion example fixtures, I've adjusted examples_test.go to
catch and compare output when opts.Run fails.