adm release: simplify `image-references` check
Previously the check for `image-references` file presence was done by a `takeFileByName` (not a very good name) function which also removed the corresponding `FileInfo` from `contents` as a sideeffect,, so it is not iterated later when the manifests are copied. The file existence can be checked directly, and the file can be easily skipped when doing substitution. While I believe the code is clearer this way, I still added a comment saying why the presence of the file is checked.