Check the node annotations only during multi-arch migration
The status command shows "The cluster is not updating" for the cases
such as MCO rolling out MachineConfigs from the cluster admins.
When the additional checks on node annotatations are introduced to
handle the multi-arch migration, we want to keep this feature intact.
So the additional checks are executed only during multi-arch migration.