pkg/cli/admin/upgrade/recommend: Initial precheck implementation
Start to set up some checks client-side, while we test conditions and
display patterns for issues with the current cluster state that may
make for bumpy updates, but which don't rise to the current
Conditional Update risk level of "we are very concerned that an update
to 4.y.z will regress on $ISSUE".  These prechecks are for issues like
"might slow an update", even if the cluster and workloads are humming
happily along except for the slowness of the update.  And other
speculative things that admins sometimes ask after, but where the link
to update speed and success is unclear.