OCPBUGS-44637: Mount proxy certificate to node-joiner pod
When a proxy is configured with a self-signed certificate, the
certificate needs to be made available to the node-joiner pod
to allow it to communicate with the proxy.

Previously, the certificate wasn't mounted to the node-joiner pod
so that when the pod attempted to pull images through the proxy, it
failed.

Now the user-ca-bundle config map is copied to the node-joiner
pod's namespace and the certificate is mounted and made available
as /etc/pki/ca-trust/extracted/pem/tls-ca-bundle.pem.