pkg/cli/admin/upgrade/recommend: Alert detail sections
Include a separate condition for each concerning alert, so cluster
admins know what we think they might want to look into.

The /%d suffixes ensure the condition types are unique, even if there
are multiple alerts with the same name firing.  Doing something more
nuanced, like only suffixing when there was a conflict, or suffixing
with something more useful than an integer, would be nice ways to
polish further in the future.