pkg/cli/admin/upgrade/recommend: New, feature-gated --accept
[1] is suggesting ClusterVersion API changes to make it easier for
cluster administrators to manage many clusters and updates vs. a more
limited set of update issues.  One example scenario:

1. A cluster admin gets conditional updates of a cluster and evaluates
   the issues in the conditional updates.
2. After evaluation, the admin decides that some issues can be accepted
   on all or some of the managed clusters.
3. The administrator runs the 'recommend' command with '--accept A,B,C
   --version 4.y.z' to see if any other issues turn up for that update
   on that cluster.  Nonzero exit code says more investigation is
   needed, some subset of which will be "additional issue D to
   assess".

Besides making life easier for admins with lots of clusters
considering one A->B update, this option will make life easier for
admins with one cluster that updates multiple hops all exposed to a
single issue (e.g. if PodDisruptionBudgetAtLimit becomes a general
issue, but the admin never wants to resolve that pre-update, they can
accept it for multiple updates in a single cluster).

The exit code handling is only tied to --version.

When listing multiple update target options, the admin is asking for
help in chosing an update target, and that cannot be bundled up into
an exit code; they'll have to read the report to get enough context to
decide.  And there may be some that are acceptable and some that have
unaccepted issues, and trying to bundle that all up into a single exit
code could go a few ways (exit non-zero if any choice has unaccepted
issues?  Exit 0 if any choice is acceptable, even if other choices have
unaccepted issues?), and it seems like folks could misinterpret
whichever approach we took and be surprised.  Exiting zero on "we
succesfully compiled this report" forces the admin to read the report
to extract any value, and that reduces the chance that they
misinterpret what we were trying to say.

[1]: https://issues.redhat.com/browse/OCPSTRAT-2118