pkg/cli/admin/upgrade/recommend: Add a --quiet option
As part of the --accepted-risk feature, because listing unaccepted
risk names gives us a way to say "we have more problems we think you
should consider" for folks who just want a short summary before
deciding whether they want more details.  I'm keeping the chatty
version as the default, because historically we have had users who
don't want to look at the details, and instead hope that waiting will
be enough for the issues to resolve.

For "there was a regression, and Red Hat will ship a fix", waiting
will eventually work.  But for "your cluster is alerting, and you
should look into that", waiting is unlikely to help.  And even for the
regression situation, it is the user's responsibility to weigh the
risk of updating into the known regression against the risk of
remaining on an old, possibly buggy release.  Red Hat will support
either choice [1].

Having --quiet be an explicit choice makes it clear that Red Hat
thinks the user should be engaging with and evaluating any detected
issues, while allowing users to say "I don't have time now, sorry.
Give me the short summary".

[1]: https://docs.redhat.com/en/documentation/openshift_container_platform/4.18/html/updating_clusters/understanding-openshift-updates-1#conditional-updates-overview_understanding-update-channels-releases