pkg/cli/admin/upgrade/recommend/alerts: Ambiguate opinions
There's a chance we might be asked to drop the
OC_ENABLE_CMD_UPGRADE_RECOMMEND_PRECHECK feature gate that's currently
in front of this logic.  If we do, this wording will make it extra
clear that we are not saying "don't update until this alert resolves".
We are saying "this alert is firing.  You should investigate and form
your own opinion".

As a concrete example, the [1] series of bugs could cause issues
pulling images, you would address the issue by launching an update to
a release with the bugfix.  If you needed to take any manual steps to
harden the core cluster components (e.g. configuring a local mirror
that was more responsive [2]), that could happen before or during the
update.

[1]: https://issues.redhat.com/browse/OCPBUGS-54658
[2]: https://docs.redhat.com/en/documentation/openshift_container_platform/4.18/html/disconnected_environments/mirroring-in-disconnected-environments