pkg/cli/admin/upgrade/recommend: Add a blank line after --version conditional risk
When I added the:

  Update to 4.18.6 Recommended=False:
  ...

block in c1ef32845c (pkg/cli/admin/upgrade/recommend: Add --version
option for specific releases, 2024-10-14, #1897), there was no
subsequent output in this "--version has conditional risks" case, so
the earlier lack-of-blank-line made sense.

But then 2d7e00420c (pkg/cli/admin/upgrade/recommend: New,
feature-gated --accept, 2025-05-06, #2023) gave the option for a
subsequent:

  error: issues that apply to this cluster but which were not included in --accept: ...

or:

  Update to %s has no known issues relevant...

This commit adds trackers to all the output blocks in the command for
"was there previous output on this stream?", so that a new block (or a
returned error) can be prefixed by the blank line needed to set off
the new block from the previous block.  For example, it now sets off
the error line from the previous, possibly-long message strings:

  ...
  Update to 4.18.6 Recommended=False:
  Image: quay.io/openshift-release-dev/ocp-release@sha256:61fdad894f035a8b192647c224faf565279518255bdbf60a91db4ee0479adaa6
  Release URL: https://access.redhat.com/errata/RHSA-2025:3066
  Reason: CRIOLayerCompressionPulls
  Message: The CRI-O container runtime may fail to pull images with certain layer compression characteristics https://issues.redhat.com/browse/OCPNODE-3074

  error: issues that apply to this cluster but which were not included in --accept: ConditionalUpdateRisk,KubeContainerWaiting

instead of cramming that 'error: ...' or '... has no known issues...'
right onto the end of the output.  The error is getting written to
standard error instead of the standard output, so there's still some
wiggle room for poor spacing for terminal users and others who are
flattening those two streams together.

And now that we have the additional newline handling in recommend.go,
I can drop the newline I'd been manually injecting in examples_test.go
to make the test fixtures more readable.  I should have noticed that
test-specific injection as a sign that the better recommend.go newline
handling was useful.  Oh well, better late than never.