Isolate manifest handling in release-extraction
There are two files `image-references`, and
 `release-metadata` that are handled differently from
 manifest files. When those files come, their readers
 from the upstream are sent to the downstream callback
 right away.

Other files contain manifests. They are parsed out
and then sent to the downstream. We will embed more
changes into this part, e.g., collecting all manifests
in the image and then use them to calculate the
enabled capabilities which is sent as an argument to
the downstream callback. Those changes are coming in
other pulls.