NO-JIRA: Include the target version in the logging
Before this full, the logging was only for the case
that `findClusterIncludeConfigFromInstallConfig` is
called, i.e., the path from an install-config
file is provided.

This pull extends it to the case where the
configuration is taken from the current cluster.

Another change from the pull is that the logging
messages include the target version that is determined
by inspecting the release image. The implementation
for this is adding a new callback `ImageConfigCallback`.