OTA-1010: extract manifests with new capabilities
The `ManifestInclusionConfiguration` determines if a manifest is
included on a cluster. Its `Capabilities` field takes the implicitly
enabled capabilities into account.

This change removes the workaround that handles the net-new capabilities
introduced by a cluster upgrade. E.g. if a cluster is currently with
4.13, then it assumes that the capabilities "build", "deploymentConfig",
and "ImageRegistry" are enabled. This is because the components
underlying those capabilities are installed by default on 4.13, or
earlier and cannot be disabled once installed. Those capabilities will
become enabled after upgrade from 4.13 to 4.14: either explicitly or
implicitly depending on the current value of
`cv.spec.capabilities.baselineCapabilitySet`.

https://github.com/openshift/oc/blob/e005223acd7c478bac070134c16f5533a258be12/pkg/cli/admin/release/extract_tools.go#L1241-L1252

CVO has already defined the function `GetImplicitlyEnabledCapabilities`
to calculate the implicitly enabled capabilities of a cluster after a
cluster upgrade. For this function to work, we have to provide

* the manifests that are currently included on the cluster, and
* the manifests from the payload in the upgrade image.

The existing `ManifestReceiver` is enhanced in a way that it can provide
enabled capabilities, including both explicit and implicit ones, when
the callback to downstream is called. It is implemented by adding a
cache to collect manifests from the upstream and calls downstream only
when all manifests are collected and the capabilities are calculated
with them using the function `GetImplicitlyEnabledCapabilities` that is
mentioned earlier.

This enhancement can be opted in by setting up the
`needEnabledCapabilities` field of `ManifestReceiver`. Otherwise, its
behaviours stays the same as before.

In case that the inclusion configuration is taken from the cluster,
i.e., `--install-config` is not set, `needEnabledCapabilities` is set to
`true`.