NO-JIRA: upgrade: use SSA for desired update
This pull replaces the function `patchDesiredUpdate` with
`applyDesiredUpdate`. The main difference is that it uses the
server-side Apply [1] via applyconfigurations [2] instead of patching
the cluster-version manifest after serializing the update.

As a result, the warning described in OCPBUGS-8356 [3] will be avoided.
Although OCPBUGS-8356 is closed as Obsolete, we will hit it again for
a new field such as `accept` [4] in `update`.

One shortcoming from this implementation is that we have to update the
code to include each new field in the future.

[1]. https://kubernetes.io/docs/reference/using-api/server-side-apply/

[2]. https://pkg.go.dev/k8s.io/client-go/applyconfigurations

[3]. https://issues.redhat.com//browse/OCPBUGS-8356

[4]. https://github.com/openshift/api/pull/2360