Add channel to detect queue is empty or not
In Observe command, we used ListKeys function to detect the DeltaFIFO is empty or not.
However, in 1.33, this function is removed entirely in DeltaFIFO. So that we need to
find a way to detect the queue is empty or not

This commit waits the execution of Pop function with a timeout. If Pop is not invoked
within the given period of time, we decide that the queue is empty and we return.