oc adm must-gather: Peform cleanup on signal
The command now removes any temporary resources on signal received,
i.e. the whole temp namespace generally, or just the pods in case
--run-namespace is specified.

To be able to implement this, kubectl logs has been pulled in
as pkg/cli/internal/logs. This is needed for now to be able to
interrupt logs streaming as LogOptions.RunLogsContext
is not yet available in v33.

We can remove kubectl logs copy in the next release
and use the relevant function, though, so added a TODO.