must-gather: Fix usage checker for custom commands
Right now the gathering process is being matched using the command string,
which doesn't work with custom commands. This change introduces a PID file,
which is used to store the running bash PID and it's being used for
sending signals to relevant processes by PID.

The checker script now uses pkill -P, which sends a signal to all child
processes of the given bash instance.