must-gather: Call queue.Done at the right time
Currently defer queue.Done(pod) is called in a worker thread loop,
which means that the pod is marked as done when the whole worker thread
exits and not in that pod processing loop iteration.

This is now fixed and the pod is marked as processed as soon as it is
actually processed.